+++ lib/src/stack_zone_specification.dart
@@ -10,9 +10,6 @@ import 'lazy_trace.dart';
 import 'trace.dart';
 import 'utils.dart';
 
-/// A function that handles errors in the zone wrapped by [Chain.capture].
-typedef _ChainHandler = void Function(dynamic error, Chain chain);
-
 /// A class encapsulating the zone specification for a [Chain.capture] zone.
 ///
 /// Until they're materialized and exposed to the user, stack chains are tracked
@@ -56,7 +53,7 @@ class StackZoneSpecification {
   ///
   /// If this is null, that indicates that any unhandled errors should be passed
   /// to the parent zone.
-  final _ChainHandler _onError;
+  final void Function(Object error, Chain) _onError;
 
   /// The most recent node of the current stack chain.
   _Node _currentNode;